+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/extensions/extensions.kt
@@ -19,7 +19,6 @@ import org.jetbrains.intellij.platform.gradle.IntelliJPlatformType
 import org.jetbrains.intellij.platform.gradle.models.IvyModule
 import org.jetbrains.intellij.platform.gradle.toIntelliJPlatformType
 import org.jetbrains.intellij.platform.gradle.utils.asPath
-import org.jetbrains.intellij.platform.gradle.utils.toVersion
 import java.io.File
 import java.nio.file.Path
 import kotlin.io.path.*
@@ -151,83 +150,3 @@ internal fun String.parsePluginNotation() = trim()
     .takeIf { it.isNotEmpty() }
     ?.split(":", "@")
     ?.run { Triple(getOrNull(0).orEmpty(), getOrNull(1).orEmpty(), getOrNull(2).orEmpty()) }
-
-// TODO: cleanup JBR helper functions:
-internal fun from(jbrVersion: String, jbrVariant: String?, jbrArch: String?, operatingSystem: OperatingSystem = OperatingSystem.current()): String {
-    val version = "8".takeIf { jbrVersion.startsWith('u') }.orEmpty() + jbrVersion
-    var prefix = getPrefix(version, jbrVariant)
-    val lastIndexOfB = version.lastIndexOf('b')
-    val lastIndexOfDash = version.lastIndexOf('-') + 1
-    val majorVersion = when (lastIndexOfB > -1) {
-        true -> version.substring(lastIndexOfDash, lastIndexOfB)
-        false -> version.substring(lastIndexOfDash)
-    }
-    val buildNumberString = when (lastIndexOfB > -1) {
-        (lastIndexOfDash == lastIndexOfB) -> version.substring(0, lastIndexOfDash - 1)
-        true -> version.substring(lastIndexOfB + 1)
-        else -> ""
-    }
-    val buildNumber = buildNumberString.toVersion()
-    val isJava8 = majorVersion.startsWith("8")
-    val isJava17 = majorVersion.startsWith("17")
-    val isJava21 = majorVersion.startsWith("21")
-
-    val oldFormat = prefix == "jbrex" || isJava8 && buildNumber < "1483.24".toVersion()
-    if (oldFormat) {
-        return "jbrex${majorVersion}b${buildNumberString}_${platform(operatingSystem)}_${arch(false)}"
-    }
-
-    val arch = jbrArch ?: arch(isJava8)
-    if (prefix.isEmpty()) {
-        prefix = when {
-            isJava17 || isJava21 -> "jbr_jcef-"
-            isJava8 -> "jbrx-"
-            operatingSystem.isMacOsX && arch == "aarch64" -> "jbr_jcef-"
-            buildNumber < "1319.6".toVersion() -> "jbr-"
-            else -> "jbr_jcef-"
-        }
-    }
-
-    return "$prefix$majorVersion-${platform(operatingSystem)}-$arch-b$buildNumberString"
-}
-
-private fun getPrefix(version: String, variant: String?) = when {
-    !variant.isNullOrEmpty() -> when (variant) {
-        "sdk" -> "jbrsdk-"
-        else -> "jbr_$variant-"
-    }
-
-    version.startsWith("jbrsdk-") -> "jbrsdk-"
-    version.startsWith("jbr_jcef-") -> "jbr_jcef-"
-    version.startsWith("jbr_dcevm-") -> "jbr_dcevm-"
-    version.startsWith("jbr_fd-") -> "jbr_fd-"
-    version.startsWith("jbr_nomod-") -> "jbr_nomod-"
-    version.startsWith("jbr-") -> "jbr-"
-    version.startsWith("jbrx-") -> "jbrx-"
-    version.startsWith("jbrex8") -> "jbrex"
-    else -> ""
-}
-
-private fun platform(operatingSystem: OperatingSystem) = when {
-    operatingSystem.isWindows -> "windows"
-    operatingSystem.isMacOsX -> "osx"
-    else -> "linux"
-}
-
-private fun arch(newFormat: Boolean): String {
-    val arch = System.getProperty("os.arch")
-    if ("aarch64" == arch || "arm64" == arch) {
-        return "aarch64"
-    }
-    if ("x86_64" == arch || "amd64" == arch) {
-        return "x64"
-    }
-    val name = System.getProperty("os.name")
-    if (name.contains("Windows") && System.getenv("ProgramFiles(x86)") != null) {
-        return "x64"
-    }
-    return when (newFormat) {
-        true -> "i586"
-        false -> "x86"
-    }
-}